extract screens
